alright now something works
Rewrote a bunch of stuff that was pointlessly complicated, added the
first example.

Initialization now works, for some reason CLKGATE was set to 0 in my PoC
but not in the "new" (this stuff has got to be a year old now, gotta
love side projects) code.

Now onto figuring where I'm at and testing and fixing and all that
stuff, maybe™.